+++ app.js
@@ -3,15 +3,14 @@
  Debugger.
 */
 
-var app, captcha, exec, express, hogan, http, lastIP, list, moment, path, port, routes, sys, template, user, word;
+var FCGI, app, captcha, exec, express, fcgi, fcgiSocket, hogan, http, lastIP, list, moment, net, path, port, routes, sys, template, user, word;
 
-process.kill(process.pid, 'SIGUSR1');
-
-sys = require('sys');
-
-exec = require('child_process').exec;
-
-exec('node-inspector');
+if (!process.argv[2]) {
+  process.kill(process.pid, 'SIGUSR1');
+  sys = require('sys');
+  exec = require('child_process').exec;
+  exec('node-inspector');
+}
 
 /*
  Module dependencies.
@@ -23,6 +22,16 @@ express = require('express');
 hogan = require('hogan.js');
 
 template = require('./routes/template');
+
+fcgi = require('connect-fastcgi');
+
+net = require('net');
+
+FCGI = require('fastcgi-stream');
+
+fcgiSocket = net.connect(7000, 'localhost', function() {
+  return console.log(JSON.stringify(arguments));
+});
 
 http = require('http');
 
@@ -75,7 +84,7 @@ app.use(function(req, res, next) {
     lastIP = req.ip;
     console.log("\n\n#####################################################");
   }
-  console.log("" + ((moment().format()).toString('yyyy-MM-dd')) + ": " + req.ip);
+  console.log("\n" + ((moment().format()).toString('yyyy-MM-dd')) + ": " + req.ip);
   return next();
 });
 
@@ -156,6 +165,33 @@ app.get('/captcha', function(req, res, next) {
 });
 
 app.post("/word/:verb", word);
+
+app.get('/fcgi', fcgi({
+  fastcgiPort: 7000,
+  fastcgiHost: 'localhost',
+  root: "./"
+}));
+
+app.get('/fcgis', function(req, res, next) {
+  var fcgiStream;
+  console.log("Sending to fcgi");
+  fcgiStream = new FCGI.FastCGIStream(fcgiSocket);
+  fcgiStream.on("record", function(requestId, record) {
+    var c, str, _i, _len, _ref;
+    console.log("RECOIRD: " + requestId + " = " + JSON.stringify(record));
+    if (record.data) {
+      str = "";
+      _ref = record.data;
+      for (_i = 0, _len = _ref.length; _i < _len; _i++) {
+        c = _ref[_i];
+        str += String.fromCharCode(c);
+      }
+      return res.send(str);
+    }
+  });
+  fcgiStream.writeRecord(6532, new FCGI.records.BeginRequest(FCGI.records.BeginRequest.roles.FILTER, FCGI.records.BeginRequest.flags.KEEP_CONN));
+  return fcgiStream.writeRecord(6532, new FCGI.records.EndRequest());
+});
 
 /*
  Start server.
